moved database connection function to own file. fixed index router and erromiddleware
