+++ app/middleware/error.middleware.ts
@@ -3,8 +3,9 @@ import Boom from 'boom';
 function errorMiddleware(error: Error, req: Request, res: Response, next: NextFunction) {
   if (error.name === 'SyntaxError') {
     res.status(400).send(Boom.badRequest().output.payload);
+  } else if (Boom.isBoom(error)) {
+    res.status(error.output.statusCode).send(error.output.payload);
   } else {
-    Boom.badRequest;
     const boom = Boom.boomify(error).output;
     res.status(boom.statusCode).send(boom.payload);
   }
